+++ .eslintrc.cjs
@@ -12,5 +12,6 @@ module.exports = {
   rules: {
     "react-refresh/only-export-components": "warn",
     "no-unused-vars": "warn",
+    "react/prop-types": "warn",
   },
 };
