+++ src/components/Navbar.jsx
@@ -1,0 +1,9 @@
+const Navbar = () => {
+  return (
+    <div>
+      <h1>Hello Navbar</h1>
+    </div>
+  );
+};
+
+export default Navbar;
