+++ src/pages/Converter.jsx
@@ -1,0 +1,9 @@
+const Converter = () => {
+  return (
+    <div>
+      <h1>Hello Converter</h1>
+    </div>
+  );
+};
+
+export default Converter;
