+++ src/pages/Currency.jsx
@@ -1,0 +1,9 @@
+const Currency = () => {
+  return (
+    <div>
+      <h1>Hello Currency</h1>
+    </div>
+  );
+};
+
+export default Currency;
