+++ src/templates/MainTemplate.jsx
@@ -1,0 +1,5 @@
+const MainTemplate = ({ children }) => {
+  return <div>{children}</div>;
+};
+
+export default MainTemplate;
